Rename mock files to correct naming
	Files renamed.
	Duplicates deleted.
