+++ src/components/connection_handler/test/include/connection_handler/mock_connection_handler_observer.h
